requirements.txt
mindsdb_native >= 2.41.0, < 2.42.0
flask-restx >= 0.2.0
flask >= 1.0
python-multipart >= 0.0.5
moz-sql-parser == 3.32.20026
pyparsing == 2.3.1
cryptography >= 2.9.2, < 3.4
pg8000 >= 1.15.3
python-tds >= 1.10.0
waitress >= 1.4.4
pymongo[tls, srv] >= 3.10.1
setuptools
psutil
sqlalchemy >= 1.3.0
sentry-sdk
walrus==0.8.2
mindsdb_datasources
flask-compress
kafka-python
appdirs
